requirements.txt
streamlit==1.38.0
beautifulsoup4==4.12.3
requests==2.32.3
pillow==10.4.0
pytesseract==0.3.13
pandas==2.2.2
